- Linux: fixed compile warnings related to %li and grt::Integer changed to ssize_t - Linux: Modeling: added Check All button to role editors [TESTS] - tested adding a role and checking/unchecking all privs

+++ backend/wbprivate/model/wb_component_physical.cpp
@@ -504,11 +504,11 @@ void WBComponentPhysical::delete_db_schema(const db_SchemaRef &schema)
     std::string objects;
 
     if (info.get_int("tables") > 0)
-      objects+= strfmt("%li tables, ", info.get_int("tables"));
+      objects+= strfmt("%li tables, ", (long)info.get_int("tables"));
     if (info.get_int("views") > 0)
-      objects+= strfmt("%li views, ", info.get_int("views"));
+      objects+= strfmt("%li views, ", (long)info.get_int("views"));
     if (info.get_int("routines") > 0)
-      objects+= strfmt("%li routines, ", info.get_int("routines"));
+      objects+= strfmt("%li routines, ", (long)info.get_int("routines"));
 
     if (!objects.empty())
       objects= objects.substr(0, objects.length()-2);
