+++ backend/wbprivate/workbench/wb_command_ui.cpp
@@ -472,7 +472,7 @@ void CommandUI::add_recent_menu(mforms::MenuItem *parent)
     std::string caption;
     if (i < 9)
     {
-      caption= strfmt("%li %s", i+1, strlist.get(i).c_str());
+      caption= strfmt("%li %s", (long)i+1, strlist.get(i).c_str());
 #if !defined(_WIN32) && !defined(__APPLE__)
       caption= "_"+replace_string(caption, "_", "__");
 #endif
@@ -493,7 +493,7 @@ void CommandUI::add_recent_menu(mforms::MenuItem *parent)
     }
     
     item = mforms::manage(new mforms::MenuItem(caption));
-    item->set_name(strfmt("wb.file.openRecentModel:%li", i+1));
+    item->set_name(strfmt("wb.file.openRecentModel:%li", (long)i+1));
     scoped_connect(item->signal_clicked(), boost::bind(&WBContext::open_recent_document, _wb, (int)i + 1));
     parent->add_item(item);
   }
